refactor(gui): extract runtime support module
